Nuanced change to Section 10 Integration Capabilities with External Systems in the CCSM requirements
Note that the requirements in the CCSM spec in section 10 referred only to a CCSM. However, depending on the CCSM, the requirements cannot be fulfilled by the CCSM and thus must be fulfilled one layer higher in the CCSM abstraction layer. Added an introduction paragraph explaining why this nuanced approach is required. Please, review and approve this change separate from section 6 on Agreement Execution.