Merge R6 and R7 (#128)
fixes #117

If there is no commercial document, R7 was meaningless and so unnecessary. While in principle an electronic version not represented in the BPI still has some utility, it doesn't seem like it's sufficient to meet the requirements of a decent system.

I didn't renumber the requirements. First because it's a massive amount of work and it isn't clear if this PR will be accepted in principle, and second because it seems possible that other requirements are changed. (I would suggest using names instead of just numbers, such as R-commdoc-in-bpi, to minimise the problem, but that's a purely editorial decision).